Move form editor components into own directory
